fix: review findings — NULL_REF sentinel, apply guard, overflow
Fix 5 confirmed bugs from code review, document 3 known limitations:

MStack: NULL_REF sentinel distinguishes "null GC ref" from "no ref".
pushRef(null) stores NULL_REF in refs[], popRef/peekRef convert back.
topIsRef() checks the raw refs[] value. Fixes null GC refs lost in
doControlTransfer, REF_IS_NULL, BR_ON_NULL, and exception catch.

Instance: apply() guard uses isObjectRef() — externref functions
now correctly throw, directing users to applyWithRefs().

Shaded: tableFillRef uses long arithmetic for overflow-safe bounds.

Compiler: null-check refArgs before loading in compileCallFunction.

InterpreterMachine: pushExceptionArgs uses tag type to determine
ref positions instead of ambiguous null check.

Known limitations (documented with TODO comments):
- Cross-module call_indirect GC ref returns discarded (Bug 4)
- Multi-value returns with Object[] vs LALOAD mismatch (Bug 5)
- Host/cross-module calls lose GC ref params in long[] (Bug 6)

Tests: GcReviewFixesTest with WAT module exercising null refs
through blocks, exceptions, table.fill bounds, and externref
via applyWithRefs.

+++ compiler/src/main/java/run/endive/compiler/internal/Shaded.java
@@ -129,7 +129,7 @@ public final class Shaded {
     public static void tableFillRef(
             int offset, Object value, int size, int tableIndex, Instance instance) {
         var table = instance.table(tableIndex);
-        if (offset + size > table.size()) {
+        if ((long) offset + (long) size > table.size()) {
             throw new TrapException("out of bounds table access");
         }
         for (int i = 0; i < size; i++) {
